refactor: Move invariant_files_mirror tests to a separate file
Moved the inline tests from src/invariant_files_mirror.rs to their own file in src/tests/invariant_files_mirror_tests.rs. This follows the pattern of other tests in the project.
