+++ src/tests/mod.rs
@@ -2,3 +2,4 @@ pub mod lru_cache_tests;
 pub mod dirty_tests;
 pub mod mirror_tests;
 pub mod block_storage_tests;
+pub mod invariant_files_mirror_tests;
